Finished CRUD operations for testResults

+++ ResultTracker/ResultTracker.API/Data/SeedData.cs
@@ -7,9 +7,14 @@ namespace ResultTracker.API.Data
 		public static void Initialise(IServiceProvider serviceProvider)
 		{
 			var context = serviceProvider.GetRequiredService<ResultTrackerDbContext>();
-			context.Results.RemoveRange(context.Results);
-			context.Subjects.RemoveRange(context.Subjects);
-			context.Topics.RemoveRange(context.Topics);
+			if (context.Topics.Any())
+			{
+				context.Results.RemoveRange(context.Results);
+				context.Subjects.RemoveRange(context.Subjects);
+				context.Topics.RemoveRange(context.Topics);
+				context.SaveChanges();
+			}
+
 
 			Topic testTopic = new Topic() { Id = Guid.Parse("d07121f2-0aa6-4a14-8d21-b086e2edf798"), Name = "Fractions", Year = "8" };
 			context.Topics.Add(testTopic);
